[Refactor, Test] Notice 도메인 작업 #113

+++ src/main/java/bleuauction/bleuauction_be/server/notice/service/NoticeService.java
@@ -1,15 +1,22 @@
 package bleuauction.bleuauction_be.server.notice.service;
 
+import bleuauction.bleuauction_be.server.attach.entity.Attach;
+import bleuauction.bleuauction_be.server.attach.service.AttachService;
+import bleuauction.bleuauction_be.server.member.entity.Member;
+import bleuauction.bleuauction_be.server.member.entity.MemberCategory;
+import bleuauction.bleuauction_be.server.ncp.NcpObjectStorageService;
 import bleuauction.bleuauction_be.server.notice.entity.Notice;
 import bleuauction.bleuauction_be.server.notice.entity.NoticeStatus;
 import bleuauction.bleuauction_be.server.notice.repository.NoticeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
+
+import static bleuauction.bleuauction_be.server.member.entity.MemberCategory.A;
 
 @Service
 @Transactional(readOnly = true)
@@ -17,11 +24,27 @@ import java.util.List;
 public class NoticeService {
 
   private final NoticeRepository noticeRepository;
+  private final NcpObjectStorageService ncpObjectStorageService;
+  private final AttachService attachService;
 
   @Transactional
-  public Long enroll(Notice notice) {
-    noticeRepository.save(notice);
-    return notice.getNoticeNo();
+  public Long enroll(Notice notice, List<MultipartFile> multipartFiles, Member member) {
+
+    if (!MemberCategory.A.equals(member.getMemberCategory())) {
+      throw new IllegalArgumentException("권한이 없습니다.");
+    }
+    notice.setMemberNo(member);
+
+    if (multipartFiles != null && !multipartFiles.isEmpty()) {
+      multipartFiles.stream()
+              .filter(file -> file.getSize() > 0)
+              .forEach(multipartFile ->
+                      notice.addNoticeAttach(ncpObjectStorageService.uploadFile(new Attach(),
+                              "bleuauction-bucket", "notice/", multipartFile))
+              );
+    }
+    return noticeRepository.save(notice).getNoticeNo();
+
   }
 
   //노티스 전체 조회
@@ -37,25 +60,52 @@ public class NoticeService {
 
   @Transactional(readOnly = true)
   public Notice findOne(Long noticeNo) {
-    return noticeRepository.findOne(noticeNo);
+    return noticeRepository.findByNoticeNo(noticeNo);
   }
 
 
   //노티스 삭제(N)
   @Transactional
-  public void deleteNotice(Long noticeNo) {
-    Notice notice = noticeRepository.findOne(noticeNo);
+  public void deleteNotice(Long noticeNo, Member member) {
+    if(!MemberCategory.A.equals(member.getMemberCategory())) {
+      throw new IllegalArgumentException("권한이 없습니다.");
+    }
+
+    Notice notice = noticeRepository.findByNoticeNo(noticeNo);
     notice.delete();
+    if (notice.getNoticeAttaches() != null && !notice.getNoticeAttaches().isEmpty()) {
+      for (Attach attach : notice.getNoticeAttaches()) {
+        attachService.changeFileStatusToDeleteByFileNo(attach.getFileNo());
+      }
+    }
+
   }
 
   //노티스 수정
   @Transactional
-  public Notice update(Notice notice) {
-    Notice updatenotice = noticeRepository.findOne(notice.getNoticeNo());
+  public Notice update(Notice updatedNotice, Member member, List<MultipartFile> multipartFiles) throws Exception{
 
-    updatenotice.setNoticeTitle(notice.getNoticeTitle());
-    updatenotice.setNoticeContent(notice.getNoticeContent());
-    return updatenotice;
+    Notice existingnotice = noticeRepository.findByNoticeNo(updatedNotice.getNoticeNo());
+
+    if(!MemberCategory.A.equals(member.getMemberCategory())) {
+      throw new IllegalArgumentException("권한이 없습니다.");
+    }
+
+      existingnotice.setNoticeTitle(updatedNotice.getNoticeTitle());
+      existingnotice.setNoticeContent(updatedNotice.getNoticeContent());
+
+    if (multipartFiles != null && !multipartFiles.isEmpty()) {
+      multipartFiles.stream()
+              .filter(file -> file.getSize() > 0)
+              .forEach(multipartFile ->
+                      existingnotice.addNoticeAttach(ncpObjectStorageService.uploadFile(new Attach(),
+                              "bleuauction-bucket", "notice/", multipartFile))
+              );
+    }
+
+      return noticeRepository.save(existingnotice);
+
+
   }
